Return concrete type for interfaces
This commit had to be made because GLib's object-oriented type system is
nowhere near as nice as Go's. I think this is inherent with any
object-oriented language/design; it's bound to be dirty and terrible and
annoying to work with. I hate it.

This change makes sense (even though it doesn't seem to with Go),
because GLib interfaces don't actually describe what a class has to
implement publicly; it only describes private methods that are made to
explicitly implement an interface.

The above design is different from Go, where interfaces describe what a
class (struct) has to implement publicly, and so its polymorphism is far
simpler: there is no need to worry about "virtual" methods, and all
methods are the same (with the exception of exporting).

Because we try to map GLib class methods to Go struct methods, we can't
really expect GLib interfaces to translate the same, so we have to
return the concrete type instead. The user can still get the underlying
type by calling Cast(), but it won't be a simple Go assertion.