Attempt to fix GC cyclic reference leaking
This commit adds a new dummy pointer field into the `intern.Box`
structure that points to a non-GC'd uintptr value (in this case the
GObject address). Using this new dummy pointer, we can attach a
finalizer to this value instead of the actual `intern.Box` value.

This is done because `intern.Box` also contains a strong reference to
the object's closure callbacks, which in turn hold a strong reference to
`intern.Box`, causing a cyclic reference.

Because the dummy pointer value points to a uintptr, it never references
anything else, so there is no GC cycle. We can then use our old
Finalizer trick to try and resurrect the `intern.Box` from the dummy
pointer and check if it should be freed in Go's heap as well.

With this, memory leaks should not happen anymore.